# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binance Fastify Service",
    "description": "Fastify service integrating Binance REST API (Market, Trading, Account, User Data Streams).",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "binanceApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MBX-APIKEY"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Liveness check",
        "tags": [
          "meta"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Readiness check",
        "tags": [
          "meta"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/market/ping": {
      "get": {
        "summary": "Test connectivity to the Rest API.",
        "tags": [
          "market"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/market/time": {
      "get": {
        "summary": "Current server time.",
        "tags": [
          "market"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/market/exchangeInfo": {
      "get": {
        "summary": "Exchange information.",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbols",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/market/depth": {
      "get": {
        "summary": "Order book (depth).",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 5,
              "maximum": 5000
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/market/trades": {
      "get": {
        "summary": "Recent trades.",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/market/klines": {
      "get": {
        "summary": "Kline/candlestick data.",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "interval",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "startTime",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "endTime",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/account/info": {
      "get": {
        "summary": "Account information (SIGNED).",
        "tags": [
          "account"
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/account/status": {
      "get": {
        "summary": "Account status (SIGNED).",
        "tags": [
          "account"
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/account/apiTradingStatus": {
      "get": {
        "summary": "API trading status (SIGNED).",
        "tags": [
          "account"
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/trade/order": {
      "post": {
        "summary": "Create a new order (SIGNED).",
        "tags": [
          "trade"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol",
                  "side",
                  "type"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ]
                  },
                  "type": {
                    "type": "string"
                  },
                  "timeInForce": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "quoteOrderQty": {
                    "type": "number"
                  },
                  "price": {
                    "type": "number"
                  },
                  "newClientOrderId": {
                    "type": "string"
                  },
                  "stopPrice": {
                    "type": "number"
                  },
                  "icebergQty": {
                    "type": "number"
                  },
                  "newOrderRespType": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "summary": "Query order (SIGNED).",
        "tags": [
          "trade"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "orderId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "origClientOrderId",
            "required": false
          }
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Cancel order (SIGNED).",
        "tags": [
          "trade"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "orderId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "origClientOrderId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "newClientOrderId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cancelRestrictions",
            "required": false
          }
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/trade/order/test": {
      "post": {
        "summary": "Test new order (SIGNED).",
        "tags": [
          "trade"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol",
                  "side",
                  "type"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ]
                  },
                  "type": {
                    "type": "string"
                  },
                  "timeInForce": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "quoteOrderQty": {
                    "type": "number"
                  },
                  "price": {
                    "type": "number"
                  },
                  "newClientOrderId": {
                    "type": "string"
                  },
                  "stopPrice": {
                    "type": "number"
                  },
                  "icebergQty": {
                    "type": "number"
                  },
                  "newOrderRespType": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/trade/openOrders": {
      "get": {
        "summary": "Current open orders (SIGNED).",
        "tags": [
          "trade"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": false
          }
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/user-stream/listenKey": {
      "post": {
        "summary": "Create a listen key (SIGNED).",
        "tags": [
          "user-stream"
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Keepalive a listen key (SIGNED).",
        "tags": [
          "user-stream"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "listenKey",
            "required": true
          }
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Close a listen key (SIGNED).",
        "tags": [
          "user-stream"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "listenKey",
            "required": true
          }
        ],
        "security": [
          {
            "binanceApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "binanceApiKey": []
    }
  ]
}